requirements.txt
python-whois==0.7.3
requests==2.32.0
